[S2GRAPH-117] moved non-play tests to s2core
