--- s2core/src/main/resources/logback.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing,
-software distributed under the License is distributed on an
-"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-KIND, either express or implied.  See the License for the
-specific language governing permissions and limitations
-under the License.
--->
-
-<configuration>
-    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-        <withJansi>true</withJansi>
-        <encoder>
-            <pattern>
-                %d{ISO8601} [%highlight(%-5level)] [%gray(%logger{0})] [%thread] - %msg%n
-            </pattern>
-        </encoder>
-    </appender>
-
-    <!--<root level="INFO">-->
-        <!--<appender-ref ref="STDOUT"/>-->
-    <!--</root>-->
-
-    <logger name="application" level="DEBUG">
-        <appender-ref ref="STDOUT"/>
-    </logger>
-
-    <logger name="error" level="DEBUG">
-        <appender-ref ref="STDOUT"/>
-    </logger>
-
-</configuration>
-
