+++ spark/build.sbt
@@ -17,16 +17,18 @@
  * under the License.
  */
 
+import Common._
+
 name := "s2spark"
 
 scalacOptions ++= Seq("-deprecation", "-feature")
 
 libraryDependencies ++= Seq(
   "com.google.guava" % "guava" % "12.0.1" force(), // use this old version of guava to avoid incompatibility
-  "org.apache.spark" %% "spark-core" % Common.sparkVersion % "provided",
-  "org.apache.spark" %% "spark-streaming" % Common.sparkVersion % "provided",
-  "org.apache.spark" %% "spark-streaming-kafka" % Common.sparkVersion,
-  "com.typesafe.play" %% "play-json" % Common.playVersion,
-  "org.specs2" %% "specs2-core" % "3.8.5" % "test",
+  "org.apache.spark" %% "spark-core" % sparkVersion % "provided",
+  "org.apache.spark" %% "spark-streaming" % sparkVersion % "provided",
+  "org.apache.spark" %% "spark-streaming-kafka" % sparkVersion,
+  "com.typesafe.play" %% "play-json" % playVersion,
+  "org.specs2" %% "specs2-core" % specs2Version % "test",
   "org.scalatest" %% "scalatest" % "2.2.1" % "test"
 )
